+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.1.0 - Unreleased
 
+* Convert to object-oriented model of configuration definition
 * Added validation tools for user-specified values for keys per [#5](https://github.com/chrispyles/fica/issues/5)
 
 ## v0.0.1 - 2022-04-17